Criado Ao_salvar_percurso_coletivo_imagens_reabertura
