+++ teste/SME.SGP.TesteIntegracao/RelatorioAcompanhamentoAprendizagem/ServicosFakes/TurmaEmPeriodoAbertoQueryHandlerFake.cs
@@ -1,0 +1,15 @@
+using System.Threading;
+using System.Threading.Tasks;
+using MediatR;
+using SME.SGP.Aplicacao;
+
+namespace SME.SGP.TesteIntegracao.RelatorioAcompanhamentoAprendizagem.ServicosFakes
+{
+    public class TurmaEmPeriodoAbertoQueryHandlerFake : IRequestHandler<TurmaEmPeriodoAbertoQuery, bool>
+    {
+        public async Task<bool> Handle(TurmaEmPeriodoAbertoQuery request, CancellationToken cancellationToken)
+        {
+            return true;
+        }
+    }
+}
